Connexion with providers/email pwd works
