+++ prisma/schema.prisma
@@ -11,45 +11,36 @@ datasource db {
 // ------- 4 "mandatory" tables better-auth -------
 
 model User {
-  id            String         @id @default(uuid())
+  id            String   @id @default(uuid())
   name          String
-  email         String         @unique
+  email         String   @unique
   emailVerified Boolean
   image         String?
-  createdAt     DateTime       @default(now())
-  updatedAt     DateTime       @updatedAt
-  sessions      Session[]      @relation("UserSessions")
-  accounts      Account[]      @relation("UserAccounts")
-  verifications Verification[] @relation("UserVerifications")
-  repoAccounts  repoAccount[]  @relation("UserRepoAccounts")
+  createdAt     DateTime @default(now())
+  updatedAt     DateTime @updatedAt
 
   @@map("user")
 }
 
 model Session {
-  id                    String    @id @default(uuid())
-  userId                String
-  user                  User      @relation("UserSessions", fields: [userId], references: [id], onDelete: Cascade)
-  expiresAt             DateTime
-  token                 String?   @unique
-  createdAt             DateTime  @default(now())
-  updatedAt             DateTime  @updatedAt
-  ipAddress             String?
-  userAgent             String?
-  accessToken           String?
-  accessTokenExpiresAt  DateTime?
-  refreshToken          String?
-  refreshTokenExpiresAt DateTime?
+  id        String   @id @default(uuid())
+  userId    String
+  token     String   @unique
+  expiresAt DateTime
+  ipAddress String?
+  userAgent String?
+  createdAt DateTime @default(now())
+  updatedAt DateTime @updatedAt
 
   @@map("session")
 }
 
+// PAS TOUCHE
 model Account {
   id                    String    @id @default(uuid())
+  userId                String
   accountId             String
   providerId            String
-  userId                String
-  user                  User      @relation("UserAccounts", fields: [userId], references: [id], onDelete: Cascade)
   accessToken           String?
   refreshToken          String?
   idToken               String?
@@ -63,13 +54,14 @@ model Account {
   @@map("account")
 }
 
+// PAS TOUCHE
 model Verification {
-  id        String   @id @default(uuid())
-  userId    String
-  user      User     @relation("UserVerifications", fields: [userId], references: [id], onDelete: Cascade)
-  expiresAt DateTime
-  createdAt DateTime @default(now())
-  updatedAt DateTime @updatedAt
+  id         String   @id @default(uuid())
+  identifier String
+  value      String
+  expiresAt  DateTime
+  createdAt  DateTime @default(now())
+  updatedAt  DateTime @updatedAt
 
   @@map("verification")
 }
@@ -81,7 +73,7 @@ model repoAccount {
   accountId             String
   providerId            String // Ex. "github", "gitlab"
   userId                String
-  user                  User      @relation("UserRepoAccounts", fields: [userId], references: [id], onDelete: Cascade)
+  // user                  User      @relation("UserRepoAccounts", fields: [userId], references: [id], onDelete: Cascade)
   accessToken           String?
   refreshToken          String?
   idToken               String?
