+++ src/lib/services/repo_account.ts
@@ -1,5 +1,5 @@
 import prisma from "@/lib/utils/prisma";
-import { auth } from "@/lib/auth/server";
+import { auth } from "@/lib/auth/auth";
 
 // Liste les comptes GitHub liés à un utilisateur.
 export async function listGitHubAccounts(req: Request) {
